Fix build process
