+++ npm/webpack-dev-server/package.json
@@ -25,9 +25,9 @@
     "webpack-dev-server": "^3.11.0"
   },
   "peerDependencies": {
-    "html-webpack-plugin": "> 3",
-    "webpack": "> 4",
-    "webpack-dev-server": "> 3"
+    "html-webpack-plugin": ">=4",
+    "webpack": ">=4",
+    "webpack-dev-server": ">=4"
   },
   "files": [
     "dist",
