+++ npm/webpack-dev-server/src/webpack.config.js
@@ -14,6 +14,6 @@ module.exports = {
     path: path.resolve(__dirname, 'dist'),
   },
   plugins: [new HtmlWebpackPlugin({
-    template: path.join(__dirname, '../../index-template.html'),
+    template: path.resolve(__dirname, '..', 'index-template.html'),
   })],
 }
